feat : 특정 시간에 대한 예약이 존재할 때 시간을 삭제하려는 경우 예외처리

+++ src/main/java/roomescape/service/ReservationTimeAdminService.java
@@ -1,9 +1,11 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.exception.ReservationExistException;
 import roomescape.exception.ReservationTimeInvalidException;
 import roomescape.model.ReservationTime;
 import roomescape.model.ReservationTimeCreateRequestDto;
+import roomescape.respository.ReservationDAO;
 import roomescape.respository.ReservationTimeDAO;
 
 import java.util.List;
@@ -13,9 +15,11 @@ import java.util.regex.Pattern;
 @Service
 public class ReservationTimeAdminService {
     private final ReservationTimeDAO reservationTimeDao;
+    private final ReservationDAO reservationDAO;
 
-    public ReservationTimeAdminService(ReservationTimeDAO reservationTimeDao) {
+    public ReservationTimeAdminService(ReservationTimeDAO reservationTimeDao, ReservationDAO reservationDAO) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDAO = reservationDAO;
     }
 
     public ReservationTime createReservationTime(ReservationTimeCreateRequestDto time) {
@@ -43,6 +47,9 @@ public class ReservationTimeAdminService {
     }
 
     public void deleteReservationTime(Long id) {
+        if (reservationDAO.isReservationExistByTimeId(id)) {
+            throw new ReservationExistException("예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteReservationTime(id);
     }
 }
